+++ common/src/client_process.rs
@@ -202,6 +202,7 @@ pub fn process_detailed(claim_data: &DataToClient, username: &String) -> DataToS
         client_version: CLIENT_VERSION.to_string(),
         unique_distribution: Some(results.distribution),
         nice_numbers: results.nice_numbers,
+        telemetry: None,
     }
 }
 
@@ -548,6 +549,7 @@ pub fn process_niceonly(claim_data: &DataToClient, username: &String) -> DataToS
         client_version: CLIENT_VERSION.to_string(),
         unique_distribution: None,
         nice_numbers: results.nice_numbers,
+        telemetry: None,
     }
 }