+++ common/src/client_process_gpu.rs
@@ -917,6 +917,7 @@ pub fn process_detailed_gpu(
         client_version: CLIENT_VERSION.to_string(),
         unique_distribution: Some(results.distribution),
         nice_numbers: results.nice_numbers,
+        telemetry: None,
     })
 }
 
@@ -937,6 +938,7 @@ pub fn process_niceonly_gpu(
         client_version: CLIENT_VERSION.to_string(),
         unique_distribution: None,
         nice_numbers: results.nice_numbers,
+        telemetry: None,
     })
 }